requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
base58==2.1.0
bitarray==1.2.2
CacheControl==0.12.6
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cytoolz==0.11.0
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
firebase-admin==5.0.3
Flask==2.0.2
google-api-core==2.1.1
google-api-python-client==2.27.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-cloud-core==2.1.0
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.41.0
hexbytes==0.2.2
httplib2==0.20.1
idna==3.3
importlib-metadata==4.8.1
ipfshttpclient==0.8.0a2
itsdangerous==2.0.1
Jinja2==3.0.2
jsonschema==3.2.0
lru-dict==1.1.7
MarkupSafe==2.0.1
msgpack==1.0.2
multiaddr==0.0.9
multidict==5.2.0
netaddr==0.8.0
packaging==21.0
parsimonious==0.8.1
proto-plus==1.19.5
protobuf==3.19.0
py-solc-x==1.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.11.0
pyparsing==2.4.7
pyrsistent==0.18.0
pywin32==302
requests==2.26.0
rlp==2.0.1
rsa==4.7.2
semantic-version==2.8.5
six==1.16.0
toolz==0.11.1
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
varint==1.0.2
web3==5.24.0
websockets==9.1
Werkzeug==2.0.2
yarl==1.7.0
zipp==3.6.0
